Restructure modules into api and core packages
