+++ leavebot/api/__init__.py
@@ -1,0 +1,2 @@
+"""API functions for LeaveBot."""
+
