+++ leavebot/api/fetch_employee.py
@@ -1,6 +1,6 @@
 import requests
 from leavebot.config.settings import EMPLOYEE_DETAILS_API, ERP_BEARER_TOKEN  # type: ignore
-from .cache_utils import EMPLOYEE_CACHE
+from ..core.cache_utils import EMPLOYEE_CACHE
 
 def fetch_employee_details(emp_id):
     """
