+++ leavebot/api/fetch_leave_balance.py
@@ -1,6 +1,6 @@
 import requests
 from leavebot.config.settings import LEAVE_SUMMARY_API, ERP_BEARER_TOKEN
-from .cache_utils import LEAVE_BALANCE_CACHE
+from ..core.cache_utils import LEAVE_BALANCE_CACHE
 
 def fetch_leave_balance(emp_id, lpd_id, from_date, to_date):
     """
