+++ leavebot/core/__init__.py
@@ -1,0 +1,2 @@
+"""Core utilities for LeaveBot."""
+
