+++ test_air_ticket_utils.py
@@ -1,5 +1,5 @@
 import unittest
-from leavebot.scripts.air_ticket_utils import next_air_ticket_eligibility
+from leavebot.core.air_ticket_utils import next_air_ticket_eligibility
 
 class TestAirTicketUtils(unittest.TestCase):
     def test_next_air_ticket_eligibility_multiple_formats(self):
